Extend UserAvatar to support orgs
